Complete User and Goal backend

+++ backend/controllers/goalController.js
@@ -2,13 +2,14 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-   const goals = await  Goal.find()
+   const goals = await  Goal.find({user: req.user.id})
 
     res.status(200).json(goals)
 })
@@ -24,7 +25,8 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(goal)
@@ -39,6 +41,19 @@ const updateGoal = asyncHandler(async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // check if logged in user and goal user are same
+    if(goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -57,6 +72,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // check if logged in user and goal user are same
+    if(goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await Goal.findByIdAndDelete(req.params.id)
